+++ test/benchee/config_test.exs
@@ -1,4 +1,0 @@
-defmodule Benchee.ConfigTest do
-  use ExUnit.Case, async: true
-  doctest Benchee.Config
-end