+++ test/benchee/configuration_test.exs
@@ -1,0 +1,4 @@
+defmodule Benchee.ConfigurationTest do
+  use ExUnit.Case, async: true
+  doctest Benchee.Configuration
+end